+++ Extension/src/background/storages/index.ts
@@ -38,3 +38,4 @@ export * from './page-stats';
 export * from './notification';
 export * from './trusted-domains';
 export * from './hit-stats';
+export * from './annoyances-consent';