+++ Extension/src/pages/options/constants.ts
@@ -72,6 +72,12 @@ export const HOW_TO_CREATE_RULES_URL = Forward.get({
     from: ForwardFrom.Options,
 });
 
+// FIXME: use the url
+export const FILTER_POLICY_URL = Forward.get({
+    action: ForwardAction.FilterPolicy,
+    from: ForwardFrom.Options,
+});
+
 export const ACCEPTABLE_ADS_LEARN_MORE_URL = Forward.get({
     action: ForwardAction.SelfPromotion,
     from: ForwardFrom.Options,